Code review and cleanup
