+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/domain/vulnerability.js
@@ -1,6 +1,6 @@
 import { UserInputError } from 'apollo-server-errors';
-import { compareValues, filterValues, updateQuery, checkIfValidUUID, validateEnumValue } from '../../utils.js';
-import { selectObjectIriByIdQuery, sanitizeInputFields } from '../../global/global-utils.js';
+import { compareValues, filterValues, updateQuery, checkIfValidUUID, validateEnumValue } from '../../../utils.js';
+import { selectObjectIriByIdQuery, sanitizeInputFields } from '../../../global/global-utils.js';
 import {
   insertVulnerabilityQuery,
   selectVulnerabilityQuery,
@@ -12,11 +12,12 @@ import {
   selectVulnerabilityQueryByIriQuery,
   attachToVulnerabilityQuery,
   detachFromVulnerabilityQuery,
-  getReducer
+  getReducer,
+  getVulnerabilityIri
 } from '../schema/sparql/vulnerability.js';
 import { createCVSSMetric, deleteCVSSMetricById } from './cvss.js';
 import { createAffectedProduct, deleteAffectedProductById } from './affectedProduct.js';
-import { createImpact, deleteImpactById } from './impact.js';
+import { createVulnerabilityImpact, deleteVulnerabilityImpactById } from './vulnerabilityImpact.js';
 import { createProblemType, deleteProblemTypeById } from './problemType.js';
 import { createTaxonomyEntry, deleteTaxonomyEntryById } from './taxonomyEntry.js';
 import { createCredit, deleteCreditById } from './credit.js';
@@ -53,9 +54,9 @@ export const createVulnerability = async (input, dbName, dataSources, select) =>
   let nestedDefinitions = {
     'metrics': { values: input.metrics, props: {}, objectType: 'metrics', createFunction: createCVSSMetric },
     'affected': { values: input.affected, props: {}, objectType: 'affected-product', createFunction: createAffectedProduct },
-    'impacts': { values: input.impacts, props: {}, objectType: 'impact-type', createFunction: createImpact },
+    'impacts': { values: input.impacts, props: {}, objectType: 'impact-type', createFunction: createVulnerabilityImpact },
     'problem_types': { values: input.problem_types, props: {}, objectType: 'problem-type', createFunction: createProblemType },
-    'taxonomy_mappings': { values: input.taxonomy_mappings, props: {}, objectType: 'taxonmy-map-entry', createFunction: createTaxonomyEntry },
+    'taxonomy_mappings': { values: input.taxonomy_mappings, props: {}, objectType: 'taxonomy-map-entry', createFunction: createTaxonomyEntry },
     'credits': { values: input.credits, props: {}, objectType: 'credit', createFunction: createCredit },
   };
 
@@ -81,7 +82,7 @@ export const createVulnerability = async (input, dbName, dataSources, select) =>
 
   // TODO: check if there are any object references
 
-  // create vulnerabilitiy
+  // create vulnerability
   let response;
   let {iri, id, query} = insertVulnerabilityQuery(input);
 
@@ -151,7 +152,7 @@ export const findVulnerabilityById = async (id, dbName, dataSources, select) => 
   // ensure the id is a valid UUID
   if (!checkIfValidUUID(id)) throw new UserInputError(`Invalid identifier: ${id}`);
 
-  let iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
+  let iri = getVulnerabilityIri(id);
   return findVulnerabilityByIri(iri, dbName, dataSources, select);
 };
 
@@ -313,7 +314,7 @@ export const deleteVulnerabilityById = async ( id, dbName, dataSources ) => {
     let nestedReferences = {
       'metrics': { object_type: 'metrics', iris: vul.metrics, deleteFunction: deleteCVSSMetricById},
       'affected': { object_type: 'affected', iris: vul.affected, deleteFunction: deleteAffectedProductById },
-      'impacts': { object_type: 'impacts', iris: vul.affected, deleteFunction: deleteImpactById },
+      'impacts': { object_type: 'impacts', iris: vul.affected, deleteFunction: deleteVulnerabilityImpactById },
       'problem_types': { object_type: 'problem_types', iris: vul.affected, deleteFunction: deleteProblemTypeById },
       'taxonomy_mappings': { object_type: 'taxonomy_mappings', iris: vul.affected, deleteFunction: deleteTaxonomyEntryById },
       'credits': { object_type: 'credits', iris: vul.affected, deleteFunction: deleteCreditById },
@@ -479,7 +480,7 @@ export const editVulnerabilityById = async (id, input, dbName, dataSources, sele
   }
 
   const query = updateQuery(
-    `http://cyio.darklight.ai/vulnerability--${id}`,
+    getVulnerabilityIri(id),
     "http://nist.gov/ns/vulnerability#Vulnerability",
     input,
     vulnerabilityPredicateMap
@@ -520,7 +521,7 @@ export const attachToVulnerability = async (id, field, entityId, dbName, dataSou
 
   // check to see if the vulnerability exists
   let select = ['id','iri','object_type'];
-  let iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
+  let iri = getVulnerabilityIri(id);
   sparqlQuery = selectVulnerabilityQueryByIriQuery(iri, select);
   
   let response;
@@ -542,7 +543,7 @@ export const attachToVulnerability = async (id, field, entityId, dbName, dataSou
     'affected': 'affected-product',
     'impacts': 'impact-type',
     'problem_types': 'problem-type',
-    'taxonomy_mappings': 'taxonmy-map-entry',
+    'taxonomy_mappings': 'taxonomy-map-entry',
     'credits': 'credit'
   };
 
@@ -570,7 +571,7 @@ export const attachToVulnerability = async (id, field, entityId, dbName, dataSou
     'affected': 'affected-product',
     'impacts': 'impact-type',
     'problem_types': 'problem-type',
-    'taxonomy_mappings': 'taxonmy-map-entry',
+    'taxonomy_mappings': 'taxonomy-map-entry',
     'credits': 'credit'
   };
 
@@ -613,7 +614,7 @@ export const detachFromVulnerability = async (id, field, entityId, dbName, dataS
 
   // check to see if the vulnerability exists
   let select = ['id','iri','object_type'];
-  let iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
+  let iri = getVulnerabilityIri(id);
   sparqlQuery = selectVulnerabilityQueryByIriQuery(iri, select);
   let response;
   try {
@@ -634,7 +635,7 @@ export const detachFromVulnerability = async (id, field, entityId, dbName, dataS
     'affected': 'affected-product',
     'impacts': 'impact-type',
     'problem_types': 'problem-type',
-    'taxonomy_mappings': 'taxonmy-map-entry',
+    'taxonomy_mappings': 'taxonomy-map-entry',
     'credits': 'credit'
   };
 
@@ -661,7 +662,7 @@ export const detachFromVulnerability = async (id, field, entityId, dbName, dataS
     'affected': 'affected-product',
     'impacts': 'impact-type',
     'problem_types': 'problem-type',
-    'taxonomy_mappings': 'taxonmy-map-entry',
+    'taxonomy_mappings': 'taxonomy-map-entry',
     'credits': 'credit'
   };
