+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/domain/vulnerabilityImpact.js
@@ -1,70 +1,98 @@
 import { UserInputError } from 'apollo-server-errors';
-import {
+import { 
   updateQuery, 
   checkIfValidUUID,
-} from '../../utils.js';
-import { sanitizeInputFields } from '../../global/global-utils.js';
+} from '../../../utils.js';
+import { sanitizeInputFields } from '../../../global/global-utils.js';
 import {
   getReducer,
-  insertCreditQuery,
-  selectCreditQuery,
-  singularizeCreditSchema,
-  deleteCreditQuery,
-  creditPredicateMap,
-  generateCreditId,
-} from '../schema/sparql/credit.js';
+  singularizeVulnerabilityImpactSchema,
+  vulnerabilityImpactPredicateMap,
+  selectVulnerabilityImpactQuery,
+  insertVulnerabilityImpactQuery,
+  deleteVulnerabilityImpactQuery,
+  selectVulnerabilityImpactByIriQuery,
+  generateVulnerabilityImpactId,
+  getVulnerabilityImpactIri,
+} from '../schema/sparql/vulnerabilityImpact.js';
 
-export const createCredit = async (input, dbName, dataSources, select) => {
+
+// TODO: Need to implment findVulnerabilityImpactByIri
+
+export const findVulnerabilityImpactByIri = async (iri, dbName, dataSources, select) => {
+  const sparqlQuery = selectVulnerabilityImpactByIriQuery(iri, select);
+  let response;
+
+  try {
+    response = await dataSources.Stardog.queryById({
+      dbName: dbName,
+      sparqlQuery,
+      queryId: "Select Vulnerability Impact",
+      singularizeSchema: singularizeVulnerabilityImpactSchema
+    });
+  } catch (e) {
+    console.log(e)
+    throw e
+  }
+
+  if (response === undefined || response === null || response.length === 0) return null;
+  const reducer = getReducer("VULNERABILITY-IMPACT");
+
+  return reducer(response[0]);  
+};
+
+export const createVulnerabilityImpact = async (input, dbName, dataSources, select) => {
   sanitizeInputFields(input);
+
   // check if impact with this same id exists
-  let checkId = generateCreditId(input);
+  let checkId = generateImpactId(input);
   // ensure the id is a valid UUID
   if (!checkIfValidUUID(checkId)) throw new UserInputError(`Invalid identifier: ${checkId}`);
 
-  // check if credit already exists
-  let selectQuery = selectCreditQuery(checkId, select);
+  let selectQuery;
+  let impact;
+  selectQuery = selectVulnerabilityImpactQuery(checkId, select);
+  try {
+    impact = await dataSources.Stardog.queryById({
+      dbName: dbName,
+      sparqlQuery: selectQuery,
+      queryId: "Select Vulnerability Impact object",
+      singularizeSchema: singularizeVulnerabilityImpactSchema
+    });
+  } catch (e) {
+    console.log(e)
+    throw e
+  }
+
+  if ( (impact != undefined && impact != null) && impact.length > 0 ) {
+    throw new UserInputError(`Cannot create vulnerabililty impact entry as entity ${checkId} already exists`);
+  }
+
+  // Insert impact object
+  let response;
+  let {iri, id, query} = insertVulnerabilityImpactQuery(input, checkId);
+  
+  try {
+    response = await dataSources.Stardog.create({
+      dbName: dbName,
+      sparqlQuery: query,
+      queryId: "Create Vulnerability Impact object"
+      });
+  } catch (e) {
+    console.log(e)
+    throw e
+  }
+  
+  // retrieve the newly created VulnerabilityImpact to be returned
+  selectQuery = selectVulnerabilityImpactQuery(id, select);
   let result;
 
   try {
     result = await dataSources.Stardog.queryById({
       dbName: dbName,
       sparqlQuery: selectQuery,
-      queryId: "Select Credit object",
-      singularizeSchema: singularizeCreditSchema
-    });
-  } catch (e) {
-    console.log(e)
-    throw e
-  }
-
-  if ( (result != undefined && result != null) && result.length > 0 ) {
-    throw new UserInputError(`Cannot create credit entry as entity ${checkId} already exists`);
-  }
-
-  let response;
-  let {iri, id, query} = insertCreditQuery(checkId, input);
-
-  try {
-    response = await dataSources.Stardog.create({
-      dbName: dbName,
-      sparqlQuery: query,
-      queryId: "Create Credit object"
-    });
-  } catch (e) {
-    console.log(e)
-    throw e
-  }
-  
-  // retrieve the newly created Credit to be returned
-  selectQuery = selectCreditQuery(id, select);
-  result;
-
-  try {
-    result = await dataSources.Stardog.queryById({
-      dbName: dbName,
-      sparqlQuery: selectQuery,
-      queryId: "Select Credit object",
-      singularizeSchema: singularizeCreditSchema
+      queryId: "Select Vulnerability Impact object",
+      singularizeSchema: singularizeVulnerabilityImpactSchema
     });
   } catch (e) {
     console.log(e)
@@ -73,23 +101,22 @@ export const createCredit = async (input, dbName, dataSources, select) => {
 
   if (result === undefined || result === null || result.length === 0) return null;
 
-  const reducer = getReducer("CREDIT");
+  const reducer = getReducer("VULNERABILITY-IMPACT");
 
-  return reducer(result[0]);
+  return reducer(result[0]); 
 };
 
-export const deleteCreditById = async ( id, dbName, dataSources ) => {
+export const deleteVulnerabilityImpactById = async ( id, dbName, dataSources ) => {
   let select = ['iri','id'];
-
   let idArray = [];
-
+  
   if (!Array.isArray(id)) {
     idArray = [id];
   } else {
     idArray = id;
   }
 
-  let removedIds = []
+  let removedIds = [];
 
   for (let itemId of idArray) {
     let response;
@@ -97,14 +124,14 @@ export const deleteCreditById = async ( id, dbName, dataSources ) => {
     if (!checkIfValidUUID(itemId)) throw new UserInputError(`Invalid identifier: ${itemId}`);  
 
     // check if object with id exists
-    let sparqlQuery = selectCreditQuery(itemId, select);
+    let sparqlQuery = selectVulnerabilityImpactQuery(itemId, select);
 
     try {
       response = await dataSources.Stardog.queryById({
         dbName,
         sparqlQuery,
-        queryId: "Select Credit object",
-        singularizeSchema: singularizeCreditSchema
+        queryId: "Select Vulnerability Impact object",
+        singularizeSchema: singularizeVulnerabilityImpactSchema
       });
     } catch (e) {
       console.log(e)
@@ -112,15 +139,14 @@ export const deleteCreditById = async ( id, dbName, dataSources ) => {
     }
 
     if (response === undefined || response.length === 0) throw new UserInputError(`Entity does not exist with ID ${itemId}`);
-    //TODO: Delete nested definitions once they are attached
 
-    sparqlQuery = deleteCreditQuery(itemId);
+    sparqlQuery = deleteVulnerabilityImpactQuery(itemId);
 
     try {
       response = await dataSources.Stardog.delete({
         dbName,
         sparqlQuery,
-        queryId: "Delete Credit"
+        queryId: "Delete Vulnerability Impact"
       });
     } catch (e) {
       console.log(e)
@@ -132,10 +158,10 @@ export const deleteCreditById = async ( id, dbName, dataSources ) => {
 
   if (!Array.isArray(id)) return id;
 
-  return removedIds; 
+  return removedIds;
 };
 
-export const editCreditById = async (id, input, dbName, dataSources, select, schema) => {
+export const editVulnerabilityImpactById = async (id, input, dbName, dataSources, select, schema) => {
   if (!checkIfValidUUID(id)) throw new UserInputError(`Invalid identifier: ${id}`);  
 
   // make sure there is input data containing what is to be edited
@@ -151,13 +177,13 @@ export const editCreditById = async (id, input, dbName, dataSources, select, sch
     editSelect.push(editItem.key);
   }
 
-  const sparqlQuery = selectCreditQuery(id, editSelect );
+  const sparqlQuery = selectVulnerabilityImpactQuery(id, editSelect );
 
   let response = await dataSources.Stardog.queryById({
     dbName,
     sparqlQuery,
-    queryId: "Select Credit",
-    singularizeSchema: singularizeCreditSchema
+    queryId: "Select Vulnerability Impact",
+    singularizeSchema: singularizeVulnerabilityImpactSchema
   });
 
   if (response.length === 0) throw new UserInputError(`Entity does not exist with ID ${id}`);
@@ -185,10 +211,10 @@ export const editCreditById = async (id, input, dbName, dataSources, select, sch
   }
 
   const query = updateQuery(
-    `http://cyio.darklight.ai/credit--${id}`,
-    "http://nist.gov/ns/vulnerability#Credit",
+    getVulnerabilityImpactIri(id),
+    "http://nist.gov/ns/vulnerability#VulnerabilityImpact",
     input,
-    creditPredicateMap
+    vulnerabilityImpactPredicateMap
   );
 
   if (query !== null) {
@@ -197,7 +223,7 @@ export const editCreditById = async (id, input, dbName, dataSources, select, sch
       response = await dataSources.Stardog.edit({
         dbName,
         sparqlQuery: query,
-        queryId: "Update Credit"
+        queryId: "Update Vulnerability Impact"
       });  
     } catch (e) {
       console.log(e)
@@ -205,38 +231,17 @@ export const editCreditById = async (id, input, dbName, dataSources, select, sch
     }
   }
 
-  const selectQuery = selectCreditQuery(id, select);
+  const selectQuery = selectVulnerabilityImpactQuery(id, select);
 
   const result = await dataSources.Stardog.queryById({
     dbName,
     sparqlQuery: selectQuery,
-    queryId: "Select Credit",
-    singularizeSchema: singularizeCreditSchema
+    queryId: "Select Vulnerability Impact",
+    singularizeSchema: singularizeVulnerabilityImpactSchema
   });
 
-  const reducer = getReducer("CREDIT");
+  const reducer = getReducer("VULNERABILITY-IMPACT");
 
   return reducer(result[0]);
 };
 
-export const findCreditByIri = async (iri, dbName, dataSources, select) => {
-  const sparqlQuery = selectCreditQuery(iri, select);
-  let response;
-
-  try {
-    response = await dataSources.Stardog.queryById({
-      dbName: dbName,
-      sparqlQuery,
-      queryId: "Select Credit",
-      singularizeSchema: singularizeCreditSchema
-    });
-  } catch (e) {
-    console.log(e)
-    throw e
-  }
-
-  if (response === undefined || response === null || response.length === 0) return null;
-  const reducer = getReducer("CREDIT");
-
-  return reducer(response[0]);  
-};
