+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/resolvers/problemType.js
