+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/resolvers/vulnerability.js
@@ -9,12 +9,12 @@ import {
 } from '../domain/vulnerability.js';
 import { findCvssMetricByIri } from '../domain/cvss.js';
 import { findAffectedProductByIri } from '../domain/affectedProduct.js';
-import { findImpactByIri } from '../domain/impact.js';
+import { findVulnerabilityImpactByIri } from '../domain/vulnerabilityImpact.js';
 import { findProblemTypesByIri } from '../domain/problemType.js';
 import { findTaxonomyByIri } from '../domain/taxonomyEntry.js';
 import { findCreditByIri } from '../domain/credit.js';
 
-import { findDataMarkingByIri } from '../../data-markings/domain/dataMarkings.js';
+import { findDataMarkingByIri } from '../../../data-markings/domain/dataMarkings.js';
 
 const cyioVulnerabilityResolvers = {
   Query: {
@@ -34,6 +34,7 @@ const cyioVulnerabilityResolvers = {
       if (parent.metrics === undefined) return [];
       let results = []
       for (let iri of parent.metrics) {
+        // TODO: this should be Metrics as there will be more than just CVSS
         let result = await findCvssMetricByIri(iri, dbName, dataSources, selectMap.getNode('metrics'));
         if (result === undefined || result === null) return null;
         results.push(result);
@@ -54,7 +55,7 @@ const cyioVulnerabilityResolvers = {
       if (parent.impacts === undefined) return [];
       let results = [];
       for (let iri of parent.impacts) {
-        let result = await findImpactByIri(iri, dbName, dataSources, selectMap.getNode('impacts'));
+        let result = await findVulnerabilityImpactByIri(iri, dbName, dataSources, selectMap.getNode('impacts'));
         if (result === undefined || result === null) return null;
         results.push(result);
       }
@@ -131,6 +132,23 @@ const cyioVulnerabilityResolvers = {
       return results;
     },
   },
+  CNATag: {
+    unsupported_when_assigned: 'unsupported-when-assigned',
+    exclusively_hosted_service: 'exclusively-hosted-service',
+    disputed: 'disputed',
+  },
+  CreditType: {
+    finder: 'finder',
+    reporter: 'reporter',
+    analyst: 'analyst',
+    coordinator: 'coordinator',
+    remediation_developer: 'remediation developer',
+    remediation_reviewer: 'remediation reviewer',
+    remediation_verifier: 'remediation verifier',
+    tool: 'tool',
+    sponsor: 'sponsor',
+    other: 'other',
+  }
 };
 
 export default cyioVulnerabilityResolvers;
