+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/resolvers/vulnerabilityImpact.js
@@ -1,0 +1,16 @@
+import {
+  createVulnerabilityImpact,
+  deleteVulnerabilityImpactById,
+  editVulnerabilityImpactById
+} from '../domain/vulnerabilityImpact.js';
+
+const cyioVulnerabilityImpactResolvers = {
+  Mutation: {
+    createVulnerabilityImpact: async (_, { input }, { dbName, dataSources, selectMap }) => createVulnerabilityImpact(input, dbName, dataSources, selectMap.getNode("createVulnerabilityImpact")), 
+    deleteVulnerabilityImpact: async (_, { id }, { dbName, dataSources }) => deleteVulnerabilityImpactById( id, dbName, dataSources),
+    deleteVulnerabilityImpacts: async (_, { ids }, { dbName, dataSources }) => deleteVulnerabilityImpactById( ids, dbName, dataSources),
+    editVulnerabilityImpact: async (_, { id, input }, { dbName, dataSources, selectMap }, {schema}) => editVulnerabilityImpactById(id, input, dbName, dataSources, selectMap.getNode("editVulnerabilityImpact"), schema),
+  }
+};
+
+export default cyioVulnerabilityImpactResolvers;
