+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/schema/sparql/taxonomyEntry.js
@@ -3,55 +3,54 @@ import {
   optionalizePredicate, 
   parameterizePredicate, 
   buildSelectVariables,
-  generateId,
+  attachQuery,
+  detachQuery,
   DARKLIGHT_NS,
-} from '../../../utils.js';
+  generateId,
+} from '../../../../utils.js';
 
 // Reducer Selection
 export function getReducer(type) {
   switch (type) {
-    case 'PROBLEMTYPE':
-      return problemTypeReducer;
+    case 'TAXONOMY-ENTRY':
+      return taxonomyEntryReducer;
     default:
       throw new UserInputError(`Unsupported reducer type ' ${type}'`)
   }
 }
 
-const problemTypeReducer = (item) => {
+const taxonomyEntryReducer = (item) => {
   // if no object type was returned, compute the type from the IRI
   if (item.object_type === undefined) {
     if (item.entity_type !== undefined) item.object_type = item.entity_type;
-    if (item.iri.includes('problem-type')) item.object_type = 'problem-type';
+    if (item.iri.includes('taxonomy-map-entry')) item.object_type = 'taxonomy-map-entry';
 }
 
 return {
     iri: item.iri,
     id: item.id,
     ...(item.object_type && { entity_type: item.object_type }),
-    ...(item.description && { description: item.description }),
-    ...(item.cwe_id && { cwe_id: item.cwe_id }),
-    ...(item.type_source && { type_source: item.type_source }),
-    ...(item.references && { references: item.references }),
+    ...(item.taxonomy_name && { taxonomy_name: item.taxonomy_name }),
+    ...(item.taxonomy_version && { taxonomy_version: item.taxonomy_version }),
+    ...(item.taxonomy_relationships && { taxonomy_relationships: item.taxonomy_relationships }),
   }
 };
 
 // Serialization schema
-export const singularizeProblemTypeSchema = { 
+export const singularizeTaxonomyEntrySchema = { 
   singularizeVariables: {
     "": false, // so there is an object as the root instead of an array
     "id": true,
     "iri": true,
     "object_type": true,
     "entity_type": true,
-    "description": true,
-    "cwe_id": true,
-    "type_source": true,
-    "references": false
+    "taxonomy_name": true,
+    "taxonomy_version": true
   }
 };
 
 // Predicate Maps
-export const problemTypePredicateMap = {
+export const taxonomyEntryPredicateMap = {
   id: {
     predicate: "<http://darklight.ai/ns/common#id>",
     binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"`: null, this.predicate, "id");},
@@ -67,47 +66,44 @@ export const problemTypePredicateMap = {
     binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "entity_type");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
-  description: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#description>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "description");},
+  taxonomy_name: {
+    predicate: "<http://csrc.nist.gov/ns/oscal/common#taxonomy_name>",
+    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "taxonomy_name");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
-  cwe_id: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#cwe_id>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "cwe_id");},
+  taxonomy_version: {
+    predicate: "<http://csrc.nist.gov/ns/oscal/common#taxonomy_version>",
+    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "taxonomy_version");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
-  type_source: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#type_source>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "type_source");},
-    optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
-  },
-  references: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#references>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "references");},
+  taxonomy_relationships: {
+    predicate: "<http://csrc.nist.gov/ns/oscal/common#taxonomy_relationships>",
+    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "taxonomy_relationships");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
 };
 
-export const generateProblemTypeId = (input) => {
-  return generateId(input, DARKLIGHT_NS);
+// Utility - TaxonomyEntry
+export const generateTaxonomyEntryId = (input) => {
+  let id_material = {...(input.taxonomy_name && {"taxonomy_name": input.taxonomy_name})} ;
+  return generateId( id_material, DARKLIGHT_NS );
 }
 
-export const insertProblemTypeQuery = (propValues) => {
-  const id = generateProblemTypeId(propValues);
+export const insertTaxonomyEntryQuery = (propValues) => {
+  const id = generateTaxonomyEntryId( propValues );
 
   // determine the appropriate ontology class type
-  const iri = `<http://cyio.darklight.ai/problem-type--${id}>`;
+  const iri = `<http://cyio.darklight.ai/taxonomy-map-entry--${id}>`;
   const insertPredicates = [];
-
+  
   Object.entries(propValues).forEach((propPair) => {
-    if (problemTypePredicateMap.hasOwnProperty(propPair[0])) {
+    if (taxonomyEntryPredicateMap.hasOwnProperty(propPair[0])) {
       if (Array.isArray(propPair[1])) {
         for (let value of propPair[1]) {
-          insertPredicates.push(problemTypePredicateMap[propPair[0]].binding(iri, value));
+          insertPredicates.push(taxonomyEntryPredicateMap[propPair[0]].binding(iri, value));
         }  
       } else {
-        insertPredicates.push(problemTypePredicateMap[propPair[0]].binding(iri, propPair[1]));
+        insertPredicates.push(taxonomyEntryPredicateMap[propPair[0]].binding(iri, propPair[1]));
       }
     }
   });
@@ -115,49 +111,51 @@ export const insertProblemTypeQuery = (propValues) => {
   const query = `
   INSERT DATA {
     GRAPH ${iri} {
-      ${iri} a <http://nist.gov/ns/vulnerability#ProblemType> .
-      ${iri} a <http://oasis-org/ns/cti/stix/domain/ProblemType> .
+      ${iri} a <http://nist.gov/ns/vulnerability#TaxonomyMapEntry> .
+      ${iri} a <http://oasis-org/ns/cti/stix/domain/TaxonomyMapEntry> .
       ${iri} a <http://darklight.ai/ns/common#Object> .
       ${iri} <http://darklight.ai/ns/common#id> "${id}" .
-      ${iri} <http://darklight.ai/ns/common#object_type> "problem-type" . 
+      ${iri} <http://darklight.ai/ns/common#object_type> "taxonomy-map-entry" . 
       ${insertPredicates.join(" . \n")}
     }
   }
   `;
 
   return {iri, id, query}
-};
+}
 
-export const selectProblemTypeQuery = (id, select) => {
-  return selectProblemTypeQueryByIriQuery(`<http://cyio.darklight.ai/problem-type--${id}>`, select);
-};
+export const selectTaxonomyEntryQuery = (id, select) => {
+  return selectTaxonomyEntryQueryByIriQuery(`<http://cyio.darklight.ai/taxonomy-map-entry--${id}>`, select);
+}
 
-export const selectProblemTypeQueryByIriQuery = (iri, select) => {
+export const selectTaxonomyEntryQueryByIriQuery = (iri, select) => {
   if (!iri.startsWith('<')) iri = `<${iri}>`;
-  if (select === undefined || select === null) select = Object.keys(problemTypePredicateMap);
+  if (select === undefined || select === null) select = Object.keys(taxonomyEntryPredicateMap);
 
   // this is needed to assist in the determination of the type of the data source
   if (!select.includes('id')) select.push('id');
   if (!select.includes('object_type')) select.push('object_type');
 
-  const { selectionClause, predicates } = buildSelectVariables(problemTypePredicateMap, select);
+  const { selectionClause, predicates } = buildSelectVariables(taxonomyEntryPredicateMap, select);
+  
   return `
   SELECT ?iri ${selectionClause}
   FROM <tag:stardog:api:context:local>
   WHERE {
     BIND(${iri} AS ?iri)
-    ?iri a <http://nist.gov/ns/vulnerability#ProblemType> .
+    ?iri a <http://nist.gov/ns/vulnerability#TaxonomyMapEntry> .
     ${predicates}
   }`
-};
+}
 
-export const deleteProblemTypeQuery = (id) => {
-  const iri = `http://cyio.darklight.ai/problem-type--${id}`;
-  return deleteProblemTypeQueryByIriQuery(iri);
-};
+export const deleteTaxonomyEntryQuery = (id) => {
+  const iri = `http://cyio.darklight.ai/taxonomy-map-entry--${id}`;
+  return deleteTaxonomyEntryByIriQuery(iri);
+}
 
-export const deleteProblemTypeQueryByIriQuery = (iri) => {
+export const deleteTaxonomyEntryByIriQuery = (iri) => {
   if (!iri.startsWith('<')) iri = `<${iri}>`;
+  
   return `
   DELETE {
     GRAPH ${iri} {
@@ -165,18 +163,18 @@ export const deleteProblemTypeQueryByIriQuery = (iri) => {
     }
   } WHERE {
     GRAPH ${iri} {
-      ?iri a <http://nist.gov/ns/vulnerability#ProblemType> .
+      ?iri a <http://nist.gov/ns/vulnerability#TaxonomyMapEntry> .
       ?iri ?p ?o
     }
   }
   `
-};
+}
 
-export const attachToProblemTypeQuery = (id, field, itemIris) => {
-  if (!problemTypePredicateMap.hasOwnProperty(field)) return null;
+export const attachToTaxonomyEntryQuery = (id, field, itemIris) => {
+  if (!taxonomyEntryPredicateMap.hasOwnProperty(field)) return null;
 
-  const iri = `<http://cyio.darklight.ai/problem-type--${id}>`;
-  const predicate = problemTypePredicateMap[field].predicate;
+  const iri = `<http://cyio.darklight.ai/taxonomy-map-entry--${id}>`;
+  const predicate = taxonomyEntryPredicateMap[field].predicate;
 
   let statements;
 
@@ -193,25 +191,23 @@ export const attachToProblemTypeQuery = (id, field, itemIris) => {
   return attachQuery(
     iri, 
     statements, 
-    problemTypePredicateMap, 
-    '<http://nist.gov/ns/vulnerability#ProblemType>'
+    taxonomyEntryPredicateMap, 
+    '<http://nist.gov/ns/vulnerability#TaxonomyMapEntry>'
   );
-};
+}
 
-export const detachFromProblemTypeQuery = (id, field, itemIris) => {
-  if (!problemTypePredicateMap.hasOwnProperty(field)) return null;
-  
-  const iri = `<http://cyio.darklight.ai/problem-type--${id}>`;
-  const predicate = problemTypePredicateMap[field].predicate;
+export const detachFromTaxonomyEntryQuery = (id, field, itemIris) => {
+  if (!taxonomyEntryPredicateMap.hasOwnProperty(field)) return null;
+
+  const iri = `<http://cyio.darklight.ai/taxonomy-map-entry--${id}>`;
+  const predicate = taxonomyEntryPredicateMap[field].predicate;
 
   let statements;
-
   if (Array.isArray(itemIris)) {
     statements = itemIris
       .map((itemIri) => `${iri} ${predicate} ${itemIri}`)
       .join(".\n        ")
-    }
-  else {
+  } else {
     if (!itemIris.startsWith('<')) itemIris = `<${itemIris}>`;
     statements = `${iri} ${predicate} ${itemIris} .`;
   }
@@ -219,7 +215,7 @@ export const detachFromProblemTypeQuery = (id, field, itemIris) => {
   return detachQuery(
     iri, 
     statements, 
-    problemTypePredicateMap, 
-    '<http://nist.gov/ns/vulnerability#ProblemType>'
+    taxonomyEntryPredicateMap, 
+    '<http://nist.gov/ns/vulnerability#TaxonomyMapEntry>'
   );
-};
+}
