+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/schema/sparql/vulnerability.js
@@ -7,7 +7,7 @@ import {
   detachQuery,
   generateId, 
   DARKLIGHT_NS,
-} from '../../../utils.js';
+} from '../../../../utils.js';
 
 // Reducer Selection
 export function getReducer(type) {
@@ -335,21 +335,28 @@ export const vulnerabilityPredicateMap = {
   },
 };
 
+
+// Utilities
 export const generateVulnerabilityId = (input) => {
-  let id_material = {...(input.vulnerability_id && {"vulnerability_id": input.vulnerability_id})} ;
+  // determine id value
+  // ${@genid}('vulnerability_id':${vulnerability_id})
+  const id_material = { ...(input.vulnerability_id && {"vulnerability_id": input.vulnerability_id}), } ;
   return generateId( id_material, DARKLIGHT_NS );
 }
 
+export const getVulnerabilityIri = (id) => {
+  if (!checkIfValidUUID(id)) throw new UserInputError(`Invalid identifier: ${id}`);
+  return `<http://cyio.darklight.ai/vulnerability--${id}>`;
+}
+
+
+// Query Builders - Vulnerability
 export const insertVulnerabilityQuery = (propValues) => {
-  // determine id value
-  // ${@genid}('vulnerability_id':${vulnerability_id})
-  const id_material = { ...(propValues.vulnerability_id && {"vulnerability_id": propValues.vulnerability_id}), } ;
-  
-  const id = generateId( id_material, DARKLIGHT_NS );
+  const id = generateVulnerabilityId(propValues);
+  const iri = getVulnerabilityIri(id);
   const timestamp = new Date().toISOString();
 
   // determine the appropriate ontology class type
-  const iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
   const insertPredicates = [];
   
   Object.entries(propValues).forEach((propPair) => {
@@ -382,7 +389,7 @@ export const insertVulnerabilityQuery = (propValues) => {
 }
 
 export const selectVulnerabilityQuery = (id, select) => {
-  return selectVulnerabilityQueryByIriQuery(`http://cyio.darklight.ai/vulnerability--${id}`, select);
+  return selectVulnerabilityQueryByIriQuery(getVulnerabilityIri(id), select);
 }
 
 export const selectVulnerabilityQueryByIriQuery = (iri, select) => {
@@ -446,7 +453,7 @@ export const selectAllVulnerabilitiesQuery = (select, args, parent) => {
 }
 
 export const deleteVulnerabilityQuery = (id) => {
-  const iri = `http://cyio.darklight.ai/vulnerability--${id}`;
+  const iri = getVulnerabilityIri(id);
   return deleteVulnerabilityByIriQuery(iri);
 }
 
@@ -469,7 +476,7 @@ export const deleteVulnerabilityByIriQuery = (iri) => {
 export const attachToVulnerabilityQuery = (id, field, itemIris) => {
   if (!vulnerabilityPredicateMap.hasOwnProperty(field)) return null;
 
-  const iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
+  const iri = getVulnerabilityIri(id);
   const predicate = vulnerabilityPredicateMap[field].predicate;
 
   let statements;
@@ -494,7 +501,7 @@ export const attachToVulnerabilityQuery = (id, field, itemIris) => {
 export const detachFromVulnerabilityQuery = (id, field, itemIris) => {
   if (!vulnerabilityPredicateMap.hasOwnProperty(field)) return null;
 
-  const iri = `<http://cyio.darklight.ai/vulnerability--${id}>`;
+  const iri = getVulnerabilityIri(id);
   const predicate = vulnerabilityPredicateMap[field].predicate;
 
   let statements;
