+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/schema/sparql/vulnerabilityImpact.js
@@ -5,51 +5,49 @@ import {
   buildSelectVariables,
   generateId,
   DARKLIGHT_NS
-} from '../../../utils.js';
+} from '../../../../utils.js';
 
 // Reducer Selection
 export function getReducer(type) {
   switch (type) {
-    case 'CREDIT':
-      return creditReducer;
+    case 'VULNERABILITY-IMPACT':
+      return vulnerabilityImpactReducer;
     default:
       throw new UserInputError(`Unsupported reducer type ' ${type}'`)
   }
 }
 
-const creditReducer = (item) => {
+const vulnerabilityImpactReducer = (item) => {
   // if no object type was returned, compute the type from the IRI
   if (item.object_type === undefined) {
     if (item.entity_type !== undefined) item.object_type = item.entity_type;
-    if (item.iri.includes('credit')) item.object_type = 'credit';
-  }
+    if (item.iri.includes('vulnerability-impact')) item.object_type = 'vulnerability-impact';
+}
 
-  return {
+return {
     iri: item.iri,
     id: item.id,
     ...(item.object_type && { entity_type: item.object_type }),
-    ...(item.credit_name && { credit_name: item.credit_name }),
-    ...(item.user_id && { user_id: item.user_id }),
-    ...(item.credit_type && { credit_type: item.credit_type }),
+    ...(item.capec_id && { capec_id: item.capec_id }),
+    ...(item.description && { description: item.description }),
   }
 };
 
 // Serialization schema
-export const singularizeCreditSchema = { 
+export const singularizeVulnerabilityImpactSchema = { 
   singularizeVariables: {
     "": false, // so there is an object as the root instead of an array
     "id": true,
     "iri": true,
     "object_type": true,
     "entity_type": true,
-    "credit_name": true,
-    "user_id": true,
-    "credit_type": true
+    "capec_id": true,
+    "description": true,
   }
 };
 
 // Predicate Maps
-export const creditPredicateMap = {
+export const vulnerabilityImpactPredicateMap = {
   id: {
     predicate: "<http://darklight.ai/ns/common#id>",
     binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"`: null, this.predicate, "id");},
@@ -65,40 +63,44 @@ export const creditPredicateMap = {
     binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "entity_type");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
-  credit_name: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#credit_name>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "credit_name");},
+  capec_id: {
+    predicate: "<http://darklight.ai/ns/common#capec_id>",
+    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"^^xsd:string` : null,  this.predicate, "capec_id");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
-  user_id: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#user_id>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "user_id");},
-    optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
-  },
-  credit_type: {
-    predicate: "<http://csrc.nist.gov/ns/oscal/common#credit_type>",
-    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"` : null,  this.predicate, "credit_type");},
+  description: {
+    predicate: "<http://darklight.ai/ns/common#description>",
+    binding: function (iri, value) { return parameterizePredicate(iri, value ? `"${value}"^^xsd:string` : null,  this.predicate, "description");},
     optional: function (iri, value) { return optionalizePredicate(this.binding(iri, value));},
   },
 };
 
-export const generateCreditId = (input) => {
+
+// Utilities
+export const generateVulnerabilityImpactId = (input) => {
   return generateId(input, DARKLIGHT_NS);
 }
 
-export const insertCreditQuery = (id, propValues) => {
+export const getVulnerabilityImpactIri = (id) => {
+  if (!checkIfValidUUID(id)) throw new UserInputError(`Invalid identifier: ${id}`); 
+  return `http://cyio.darklight.ai/vulnerability-impact--${id}`; 
+}
+
+
+
+export const insertVulnerabilityImpactQuery = (id, propValues) => {
   // determine the appropriate ontology class type
-  const iri = `<http://cyio.darklight.ai/credit--${id}>`;
+  const iri = getVulnerabilityImpactIri(id);
   const insertPredicates = [];
   
   Object.entries(propValues).forEach((propPair) => {
-    if (creditPredicateMap.hasOwnProperty(propPair[0])) {
+    if (vulnerabilityImpactPredicateMap.hasOwnProperty(propPair[0])) {
       if (Array.isArray(propPair[1])) {
         for (let value of propPair[1]) {
-          insertPredicates.push(creditPredicateMap[propPair[0]].binding(iri, value));
+          insertPredicates.push(vulnerabilityImpactPredicateMap[propPair[0]].binding(iri, value));
         }  
       } else {
-        insertPredicates.push(creditPredicateMap[propPair[0]].binding(iri, propPair[1]));
+        insertPredicates.push(vulnerabilityImpactPredicateMap[propPair[0]].binding(iri, propPair[1]));
       }
     }
   });
@@ -106,11 +108,10 @@ export const insertCreditQuery = (id, propValues) => {
   const query = `
   INSERT DATA {
     GRAPH ${iri} {
-      ${iri} a <http://nist.gov/ns/vulnerability#Credit> .
-      ${iri} a <http://oasis-org/ns/cti/stix/domain/Credit> .
-      ${iri} a <http://darklight.ai/ns/common#Object> .
+      ${iri} a <http://nist.gov/ns/vulnerability#VulnerabilityImpact> .
+      ${iri} a <http://darklight.ai/ns/common#ComplexDatatype> .
       ${iri} <http://darklight.ai/ns/common#id> "${id}" .
-      ${iri} <http://darklight.ai/ns/common#object_type> "credit" . 
+      ${iri} <http://darklight.ai/ns/common#object_type> "vulnerability-impact" . 
       ${insertPredicates.join(" . \n")}
     }
   }
@@ -119,38 +120,38 @@ export const insertCreditQuery = (id, propValues) => {
   return {iri, id, query}
 }
 
-export const selectCreditQuery = (id, select) => {
-  return selectCreditQueryByIriQuery(`<http://cyio.darklight.ai/credit--${id}>`, select);
-};
+export const selectVulnerabilityImpactQuery = (id, select) => {
+  return selectVulnerabilityImpactByIriQuery(getVulnerabilityImpactIri(id), select);
+}
 
-export const selectCreditQueryByIriQuery = (iri, select) => {
+export const selectVulnerabilityImpactByIriQuery = (iri, select) => {
   if (!iri.startsWith('<')) iri = `<${iri}>`;
-  if (select === undefined || select === null) select = Object.keys(creditPredicateMap);
+  if (select === undefined || select === null) select = Object.keys(vulnerabilityImpactPredicateMap);
 
   // this is needed to assist in the determination of the type of the data source
   if (!select.includes('id')) select.push('id');
   if (!select.includes('object_type')) select.push('object_type');
 
-  const { selectionClause, predicates } = buildSelectVariables(creditPredicateMap, select);
-  
+  const { selectionClause, predicates } = buildSelectVariables(vulnerabilityImpactPredicateMap, select);
+
   return `
   SELECT ?iri ${selectionClause}
   FROM <tag:stardog:api:context:local>
   WHERE {
     BIND(${iri} AS ?iri)
-    ?iri a <http://nist.gov/ns/vulnerability#Credit> .
+    ?iri a <http://nist.gov/ns/vulnerability#VulnerabilityImpact> .
     ${predicates}
   }`
-};
+}
 
-export const deleteCreditQuery = (id) => {
-  const iri = `http://cyio.darklight.ai/credit--${id}`;
-  return deleteCreditByIriQuery(iri);
-};
+export const deleteVulnerabilityImpactQuery = (id) => {
+  const iri = getVulnerabilityImpactIri(id);
+  return deleteVulnerabilityImpactByIriQuery(iri);
+}
 
-export const deleteCreditByIriQuery = (iri) => {
+export const deleteVulnerabilityImpactByIriQuery = (iri) => {
   if (!iri.startsWith('<')) iri = `<${iri}>`;
-
+  
   return `
   DELETE {
     GRAPH ${iri} {
@@ -158,9 +159,9 @@ export const deleteCreditByIriQuery = (iri) => {
     }
   } WHERE {
     GRAPH ${iri} {
-      ?iri a <http://nist.gov/ns/vulnerability#Credit> .
+      ?iri a <http://nist.gov/ns/vulnerability#VulnerabilityImpact> .
       ?iri ?p ?o
     }
   }
   `
-};
+}
