+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/typeDefs/cvss.graphql
@@ -19,7 +19,6 @@ extend type Mutation {
   editCVSSMetric(id: ID!, input: [EditInput]!, commitMessage: String): CVSSMetric @hasPermission(resources: "Context:update")
   attachToCVSSMetric(id: ID!, field: String!, entityId: ID!): Boolean @hasPermission(resources: "Context:update")
   detachFromCVSSMetric(id: ID!, field: String!, entityId: ID!): Boolean @hasPermission(resources: "Context:update")
-
 }
 
 interface CVSS {
