+++ opencti-platform/opencti-graphql/src/cyio/schema/threat-intelligence/domain-objects/typeDefs/vulnerability.graphql
@@ -31,11 +31,11 @@ extend type Mutation {
   deleteCredit(id: ID!): ID @hasPermission(resources: "Context:delete")
   deleteCredits(ids: [ID!]!): [ID] @hasPermission(resources: "Context:delete")
   editCredit(id: ID!, input: [EditInput]!, commitMessage: String): Credit @hasPermission(resources: "Context:update")
-  # ImpactType
-  createImpact(input: ImpactTypeInput!): ImpactType @hasPermission(resources: "Context:update")
-  deleteImpact(id: ID!): ID @hasPermission(resources: "Context:delete")
-  deleteImpacts(ids: [ID!]!): [ID] @hasPermission(resources: "Context:delete")
-  editImpact(id: ID!, input: [EditInput]!, commitMessage: String): ImpactType @hasPermission(resources: "Context:update")
+  # VulnerabilityImpact
+  createVulnerabilityImpact(input: VulnerabilityImpactInput!): VulnerabilityImpact @hasPermission(resources: "Context:update")
+  deleteVulnerabilityImpact(id: ID!): ID @hasPermission(resources: "Context:delete")
+  deleteVulnerabilityImpacts(ids: [ID!]!): [ID] @hasPermission(resources: "Context:delete")
+  editVulnerabilityImpact(id: ID!, input: [EditInput]!, commitMessage: String): VulnerabilityImpact @hasPermission(resources: "Context:update")
   # MetricType
   createMetric(input: MetricTypeInput!): MetricType @hasPermission(resources: "Context:update")
   deleteMetric(id: ID!): ID @hasPermission(resources: "Context:delete")
@@ -142,7 +142,7 @@ type Vulnerability implements BasicObject & LifecycleObject & CoreObject {
   "Indicates the list of information for each problem type."
   problem_types: [ProblemType]
   "Indicates the collection of impacts of this vulnerability."
-  impacts: [ImpactType]
+  impacts: [VulnerabilityImpact]
   "Indicates the collection of impact scores with attribution."
   metrics: [MetricType!]
   "Indicates the configurations required for exploiting this vulnerability."
@@ -210,7 +210,7 @@ input VulnerabilityInput {
   "Indicates the list of information for each problem type."
   problem_types: [ProblemTypeInput]
   "Indicates the collection of impacts of this vulnerability."
-  impacts: [ImpactTypeInput]
+  impacts: [VulnerabilityImpactInput]
   "Indicates the collection of impact scores with attribution."
   metrics: [MetricTypeInput]
   "Indicates the configurations required for exploiting this vulnerability."
@@ -392,20 +392,20 @@ input CreditInput {
 """
 Provides information about the impact from this vulnerability.
 """
-type ImpactType implements ComplexDatatype {
+type VulnerabilityImpact implements ComplexDatatype {
   # ComplexDatatype
   "Uniquely identifies this object."
   id: ID!
   "Identifies the type of the Object."
   entity_type: String!
-  # Impact Type
+  # Vulnerability Impact
   "Indicates the CAPEC ID that best relates to this impact."
   capec_id: String
   "Indicates a 1-4096 characters of text in a particular language with optional alternate markup or formatted representation (e.g., Markdown)."
   description: String
 }
-input ImpactTypeInput {
-  # Impact Type
+input VulnerabilityImpactInput {
+  # Vulnerability Impact
   "Indicates the CAPEC ID that best relates to this impact."
   capec_id: String
   "Indicates a 1-4096 characters of text in a particular language with optional alternate markup or formatted representation (e.g., Markdown)."
